correction views + draft functionnalitites for news & fics

+++ src/View/News/NewsView.php
@@ -4,7 +4,11 @@ $_SESSION['page'] = "News";
 
 include(__DIR__."/../../Model/News/NewsModel.php");
 $newsModel = new NewsModel();
-$news = $newsModel->getNews();
+if(isset($_SESSION['role']) && $_SESSION['role']=='admin'){
+    $news = $newsModel->getAllNews();
+} else {
+    $news = $newsModel->getNews();
+}
 unset($newsModel);
 
 ?>
@@ -16,30 +20,34 @@ unset($newsModel);
     <div id="page">
         <?php include(__DIR__."/../nav.php"); ?>
 
-        <div style="margin-left: 20px; margin-top: 10px;"><button>Trier par</button><?php if(isset($_SESSION['role']) && $_SESSION['role']=='admin'):?><a href="/src/View/News/NewsAdd.php">Publier une news</a><?php endif;?></div>
+        <main>
+            <?php if(isset($_SESSION['role']) && $_SESSION['role']=='admin'):?><a href="/src/View/News/NewsAdd.php">Publier une news</a><?php endif;?>
 
+            <div class="grid-y align-spaced callout">
 
-        <div class="grid-y align-spaced callout news-container">
-
-            <?php foreach($news as $new): ?>
-                    <div class='grid-y align-spaced callout small news-news-container'>
-                        <div class='grid-x align-justify'>
-                            <div class='news-title-container'><h2><?= $new[1];?></h2></div>
-                            <div class='news-author-container'><?php if(isset($_SESSION['role']) && $_SESSION['role']=='admin'):?>
-                                    <a href="/src/View/News/NewsEdit.php?id=<?=$new[0];?>">Modifier</a>
-                                    <a href="/src/View/News/NewsDelete.php?id=<?=$new[0];?>">Supprimer</a>
-                                    <?php endif;?><?= $new[3]." - <a href='/src/View/Account/AccountView.php?id=".$new[5]."'>".$new[4]."</a>";?></div>
+                <?php foreach($news as $new): ?>
+                        <div class="grid-y align-spaced callout small news-news-container">
+                            <div class="grid-x align-justify">
+                                <div><h2><?= $new[1];?></h2></div>
+                                <div>
+                                    <?php if(isset($_SESSION['role']) && $_SESSION['role']=='admin'):?>
+                                        <a href="/src/View/News/NewsEdit.php?id=<?=$new[0];?>">Modifier</a>
+                                        <a href="/src/View/News/NewsDelete.php?id=<?=$new[0];?>">Supprimer</a>
+                                    <?php endif;?>
+                                    <?php
+                                    if($new[6]==1) echo $new[3];
+                                    else echo "Brouillon";?>
+                                    <a href="/src/View/Account/AccountView.php?id=<?=$new[5];?>"><?=$new[4];?></a></div>
+                            </div>
+                            <div class="cell auto news-message-container"><p><?= $new[2];?></p></div>
                         </div>
-                        <div class='cell auto news-message-container'><p><?= $new[2];?></p></div>
-                    </div>
-            <?php endforeach; ?>
+                <?php endforeach; ?>
 
-        </div>
+            </div>
+        </main>
 
         <?php include(__DIR__."/../footer.php"); ?>
     </div>
 
 </body>
 
-
-</html>
